moves EditDataBox and ViewDataTable from DataTable to DataPanel folder as these make up their own logic on how new data is entered and should be kept separate from the previous logic
