+++ src/login/components/DataPanel/DataPanel.js
@@ -1,8 +1,8 @@
 import React, { Component, Fragment } from "react";
 import PropTypes from "prop-types";
 import InjectIntl from "../../translation/InjectIntl_HOC_factory";
-import ViewDataTable from "../DataTable/ViewDataTable";
-import EditDataBox from "../DataTable/EditDataBox";
+import ViewDataTable from "./ViewDataTable";
+import EditDataBox from "./EditDataBox";
 
 const RenderViewData = (props) => {
   return (
